+++ pkg/control/controldisplay/get_exporters.go
@@ -1,12 +1,14 @@
 package controldisplay
 
 import (
+	"context"
+
 	"github.com/turbot/steampipe/pkg/export"
 )
 
 // GetExporters returns an array of ControlExporters corresponding to the available output formats
-func GetExporters() ([]export.Exporter, error) {
-	formatResolver, err := NewFormatResolver()
+func GetExporters(ctx context.Context) ([]export.Exporter, error) {
+	formatResolver, err := NewFormatResolver(ctx)
 	if err != nil {
 		return nil, err
 	}