+++ pkg/initialisation/init_data.go
@@ -49,9 +49,12 @@ func NewInitData() *InitData {
 
 func (i *InitData) RegisterExporters(exporters ...export.Exporter) *InitData {
 	for _, e := range exporters {
-		i.ExportManager.Register(e)
+		if err := i.ExportManager.Register(e); err != nil {
+			// short circuit if there is an error
+			i.Result.Error = err
+			return i
+		}
 	}
-
 	return i
 }
 
@@ -112,7 +115,7 @@ func (i *InitData) Init(ctx context.Context, invoker constants.Invoker, opts ...
 
 	statushooks.SetStatus(ctx, "Checking for required plugins")
 	log.Printf("[INFO] Checking for required plugins")
-	pluginsInstalled, err := plugin.GetInstalledPlugins()
+	pluginsInstalled, err := plugin.GetInstalledPlugins(ctx)
 	if err != nil {
 		i.Result.Error = err
 		return